streamlit/pyproject.toml
[project]
name = "manufacturing-wiki"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "snowflake-connector-python>=3.3.0",
    "streamlit[snowflake]>=1.41.0",
    "pandas>=2.0.0",
]
